Disable sprint for now
Will need a cache setup with sprint animations for both elite and spartan fp weapons with a HUD element to match, having these be missing results in unfair game-play. While 90% of this can be achieved quite easily by borrowing from 0.6 the elite animations would need to be converted from ms25.